Adds `@types/chrome` dependency.
Refactored to use a "solution-style" `tsconfig` so type dependencies are explicitly listed in `tsconfig.sw.json` while they are disabled from the root `tsconfig.base.json`. This forces each TS compilation to explicitly state which types are available in its environment. This required a solution-style config in order work with IDEs.