Disables caching.
Somehow the extension seems to have cached the result of `/index.ts` and refuses to drop or ignore it no matter what I do. Updated the server to disable caching altogether, but also had to set the `cache` argument in `fetch()` to get the extension to correctly update. No idea what's going on here TBH.