Serve `node_modules/` and skip transformation for files from it.
This is a consequence of the inelegant `*.js` -> `*.ts` redirect, since we need to skip this for pre-built dependencies with no good way of knowing which resources are coming from NPM. The simple workaround is to bind to the `node_modules/` name, and exempt any URL which is coming from that directory. In practice, something like Vite is probably the better answer here, but if you're using Vite then you can already easily transform TS to JS anyways, so there's not much value in this extension.